refactor: inline theme organiser helpers
